Issues fixed

+++ app/Http/Middleware/IsInstalled.php
@@ -32,7 +32,10 @@ class IsInstalled
 
     public function alreadyInstalled()
     {
+        if(file_exists(storage_path('installed')))
+        {
         unlink(storage_path('installed'));
+        }
 
         return file_exists(storage_path('installed'));
     }
